Rename the services mixin
