+++ lib/torii/services/iframe.js
@@ -1,8 +1,8 @@
-import ServicesMixin from 'torii/mixins/services-mixin';
+import UiServiceMixin from 'torii/mixins/ui-service-mixin';
 
 var on = Ember.on;
 
-var Iframe = Ember.Object.extend(Ember.Evented, ServicesMixin, {
+var Iframe = Ember.Object.extend(Ember.Evented, UiServiceMixin, {
 
   // Open an iframe window.
   openImpl : function(url, pendingRequestKey, optionsString){
