+++ lib/torii/services/popup.js
@@ -1,8 +1,8 @@
-import ServicesMixin from 'torii/mixins/services-mixin';
+import UiServiceMixin from 'torii/mixins/ui-service-mixin';
 
 var on = Ember.on;
 
-var Popup = Ember.Object.extend(Ember.Evented, ServicesMixin, {
+var Popup = Ember.Object.extend(Ember.Evented, UiServiceMixin, {
 
   // Open a popup window.
   openImpl : function(url, pendingRequestKey, optionsString){
